auth/requirements.txt
asyncpg

fastapi
pyjwt
argon2-cffi
pydantic[email]
itsdangerous
aiosmtplib
uvicorn[standard]